# docs/API_SPEC.yml
openapi: 3.1.0
info:
  title: MiTechoRentable API
  version: 0.1.0
  description: |
    **FastAPI = servicio aislado** (cálculo + borde de ingesta). **Supabase** concentra Auth (GoTrue),
    JWT, RLS y Postgres. Este spec no define login: el cliente React usa `@supabase/supabase-js` **≥ 2.70**
    (`getSession()`, `access_token`) y envía ese token a las rutas user-facing como `Authorization: Bearer`.

    **Validación JWT en FastAPI:** misma clave/secreto que el proyecto Supabase (Settings → API → JWT).
    Revisar `aud`, `iss`, `exp` según versión del proyecto; alinear con la guía oficial de la versión **2.70.x**
    del cliente JS que usen en frontend.

    **Obtención de datos (pipeline):**
    1. Dispositivos / APIs de marca → **middleware Tinku** (`docs/resources/technical_guide.md`, API key equipo).
       Contrato y rutas usadas hacia `techos.thetribu.dev`: **`docs/TINKU_MIDDLEWARE_API.yml`** (manifest ejecutable en
       `backend/tests/fixtures/tinku_endpoints.yaml`).
    2. Script o worker agrega lecturas en `IngestBatch` → **POST /ingest** con **X-Ingest-Key** (máquina a máquina,
       no JWT de usuario). FastAPI infiere magnitudes y persiste vía `service_role` en servidor.
    3. Usuarios humanos → React lee Postgres con supabase-js (RLS) y llama a GET/WS de esta API con **JWT Supabase**.

    Rutas protegidas con JWT exigen que el `sub` del token tenga permiso sobre el recurso (reflejado en claims
    o comprobación en servidor tras validar firma).

...
servers:
...
paths:
...
  /stats/{client_id}:
    get:
      tags: [stats]
      summary: KPIs financieros, ambientales y técnicos por cliente
      security:
        - supabaseJWT: []
      parameters:
        - in: path
          name: client_id
          required: true
          schema: { type: string, format: uuid }
        - in: query
          name: from
          schema: { type: string, format: date-time }
        - in: query
          name: to
          schema: { type: string, format: date-time }
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientStats"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /reports/generate:
    get:
      tags: [reports]
      summary: Generación on-demand de reporte PDF
      security:
        - supabaseJWT: []
      parameters:
        - in: query
          name: client_id
          required: true
          schema: { type: string, format: uuid }
        - in: query
          name: period
          required: true
          schema: { type: string, enum: [month, quarter, year, custom] }
        - in: query
          name: from
          schema: { type: string, format: date-time }
        - in: query
          name: to
          schema: { type: string, format: date-time }
      responses:
        "200":
          description: PDF binario
          content:
            application/pdf:
              schema: { type: string, format: binary }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
components:
  securitySchemes:
    supabaseJWT:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        **Access token** emitido por Supabase Auth (mismo que entrega `getSession()` en JS 2.70+).
        FastAPI valida firma con el JWT secret del proyecto; no reimplementa login.
    ingestApiKey:
      type: apiKey
      in: header
      name: X-Ingest-Key
      description: |
        Secreto compartido entre middleware Tinku (o worker) y FastAPI. Rotar en producción; nunca en el frontend.
  responses:
    BadRequest:
      description: request inválido
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: token/clave inválidos o expirados
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: recurso no encontrado
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Error:
      type: object
      required: [detail]
      properties:
        detail: { type: string }
        code: { type: string }
...
    ClientStats:
      type: object
      properties:
        client_id: { type: string, format: uuid }
        window:
          type: object
          properties:
            from: { type: string, format: date-time }
            to: { type: string, format: date-time }
        financial:
          type: object
          properties:
            savings_accumulated_usd: { type: number }
            roi_curve:
              type: array
              items:
                type: object
                properties:
                  t: { type: string, format: date-time }
                  cumulative_savings_usd: { type: number }
        environmental:
          type: object
          properties:
            co2_avoided_kg: { type: number }
            trees_equivalent: { type: number }
        technical:
          type: object
          properties:
            energy_kwh: { type: number }
            uptime_contractual_pct: { type: number }
            power_factor_mean: { type: number }
            degradation_trend_pct_per_year: { type: number }
...